fix(stacktrace): Fix frame registers not aligning to grid or being truncated by ClippedBox (#60023)
Fixes https://github.com/getsentry/sentry/issues/60013

I believe this regression is due to the refactor of ClippedBox here:
https://github.com/getsentry/sentry/pull/59524

This component was adding a lot of custom behavior to the ClippedBox
component which seemed to break after the refactor. IMO those
customizations did not add enough value to be worth the inconsistency
and additional maintenance.