feat(database): Adaptive `interval` for duration charts (#60278)
Preparation for releasing percentiles. Percentiles have much worse
performance than counters, so we need to reduce the granularity that we
fetch by reducing the `interval` we ask for. _However_ I don't want to
sacrifice the high granularity we get for simple counter metrics like
`avg`. This PR sets the `interval` query parameter based on the `yAxis`
value. If the `yAxis` includes any distribution metrics, it turns the
granularity down.

I had to add some helpers and specs along the way to support this.

- `GranularityLadder` is an array object that specifies granularities
for minute intervals. Check out `COUNTER_GRANULARITIES` for an example.
This is way easier and cleaner to tweak than constantly updating the
`getInterval` function
- Add `findGranularityIntervalForMinutes` helper function to complement
the `GranularityLadder` type
- Declare correct granularities for Database metrics
- Dynamically adjust granularity based on metric series