test(hc): Change default test behavior to silo modes (#60426)
Have all test cases tagged with a silo mode decorator
(`@control_silo_test` or `@region_silo_test`) run only in that mode. The
`stable` argument is now true by default, and the test is not run in
monolith mode unless it has `include_monolith_run=True`.

Repeating #60291, which was reverted due to a `getsentry` test failure.